Revert "PJKT-26 minor changes for file project naming and package structure"
This reverts commit 587d1cbe65763257d9ac1b6e6e7932837d3114f5.
