+++ pom.xml
@@ -8,10 +8,11 @@
 		<version>3.4.4</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.github.StudentsDreamTeam</groupId>
-	<artifactId>best-in-quest-api</artifactId>
+	<groupId>com.example</groupId>
+	<artifactId>demo</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>best-in-quest-api</name>
+	<name>demo</name>
+	<description>Demo project for Spring Boot</description>
 	<url/>
 	<licenses>
 		<license/>
@@ -26,7 +27,7 @@
 		<url/>
 	</scm>
 	<properties>
-		<java.version>24</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -48,7 +49,7 @@
             <artifactId>spring-boot-maven-plugin</artifactId>
             <version>3.4.4</version>
             <configuration>
-                <mainClass>com.github.StudentsDreamTeam.RestServiceApplication</mainClass>
+                <mainClass>com.example.restservice.RestServiceApplication</mainClass>
             </configuration>
         </plugin>
     </plugins>
