+++ src/main/java/com/example/demo/DemoApplication.java
@@ -1,13 +1,13 @@
-package com.github.StudentsDreamTeam;
+package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class RestServiceApplication {
+public class DemoApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(RestServiceApplication.class, args);
+		SpringApplication.run(DemoApplication.class, args);
 	}
 
 }
