+++ src/main/java/com/example/restservice/Greeting.java
@@ -1,3 +1,3 @@
-package com.github.StudentsDreamTeam;
+package com.example.restservice;
 
 public record Greeting(long id, String content) { }
