+++ src/main/java/com/example/restservice/GreetingController.java
@@ -1,0 +1,19 @@
+package com.example.restservice;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class GreetingController {
+
+	private static final String template = "Hello, %s!";
+	private final AtomicLong counter = new AtomicLong();
+
+	@GetMapping("/greeting")
+	public Greeting greeting(@RequestParam(value = "BestInQuest", defaultValue = "Ilya") String name) {
+		return new Greeting(counter.incrementAndGet(), String.format(template, name));
+	}
+}
