+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=best-in-quest-api
+spring.application.name=demo
